Add files via upload
Credit Card Fraud Detection with Random Forest Algorithm

The Random Forest is an ensemble learning method that operates by constructing multiple decision trees during training and outputting the class that is the mode of the classes of the individual trees. It is particularly effective in classification tasks, such as fraud detection, due to its robustness against overfitting and its ability to handle imbalanced datasets.